refac(dot): move user.dto.ts from auth/ to users/
