+++ apps/backend/src/delivery/http/auth/auth.controller.ts
@@ -11,10 +11,10 @@ import { RequestChangeEmailDto } from "@/services/auth/dto/requestChangeEmail.dt
 import { ResetPasswordByTokenDto } from "@/services/auth/dto/resetPasswordByToken.dto"
 import { SignInDto } from "@/services/auth/dto/signIn.dto"
 import { SignUpDto } from "@/services/auth/dto/signUp.dto"
-import { UserDto } from "@/services/auth/dto/user.dto"
 import { AuthGuard } from "@/services/auth/guards/auth.guard"
 import { AuthProviderGuard } from "@/services/auth/guards/provider.guard"
 import { ProvidersService } from "@/services/auth/providers/providers.service"
+import { UserDto } from "@/services/users/dto/user.dto"
 import { User } from "@/services/users/user.decorator"
 import { AlreadyAuthenticatedError, NoCodeProvidedError, SessionError } from "@/shared/errors/auth"
 import { getMillisecondsFromHours } from "@/shared/utils/getMillisecondsFromHours"
