+++ apps/backend/src/delivery/http/mediaItems/mediaItems.controller.ts
@@ -1,6 +1,5 @@
 import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UseGuards } from "@nestjs/common"
 import { isCuid } from "@paralleldrive/cuid2"
-import { UserDto } from "@/services/auth/dto/user.dto"
 import { AuthGuard } from "@/services/auth/guards/auth.guard"
 import { BulkCreateMediaItemDto } from "@/services/mediaItems/dto/bulkCreateMediaItem.dto"
 import { BulkDeleteMediaItemDto } from "@/services/mediaItems/dto/bulkDeleteMediaItem.dto"
@@ -11,6 +10,7 @@ import { GetMediaItemsByMediaIdParams } from "@/services/mediaItems/dto/getMedia
 import { GetMediaItemsCountByListIdQueryDto } from "@/services/mediaItems/dto/getMediaItemsCountByListIdQuery.dto"
 import { UpdateMediaItemDto } from "@/services/mediaItems/dto/updateMediaItem.dto"
 import { MediaItemsService } from "@/services/mediaItems/mediaItems.service"
+import { UserDto } from "@/services/users/dto/user.dto"
 import { User } from "@/services/users/user.decorator"
 import { MediaItemListIdDto } from "@/shared/dto/mediaItemListId.dto"
 import { UuidDto } from "@/shared/dto/uuid.dto"
