+++ apps/backend/src/delivery/http/mediaListViews/mediaListViews.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Param, Post, UseGuards } from "@nestjs/common"
-import { UserDto } from "@/services/auth/dto/user.dto"
 import { AuthGuard } from "@/services/auth/guards/auth.guard"
 import { MediaListViewsService } from "@/services/mediaListViews/mediaListViews.service"
+import { UserDto } from "@/services/users/dto/user.dto"
 import { User } from "@/services/users/user.decorator"
 
 @Controller("media-list-views")
