+++ apps/backend/src/delivery/http/mediaLists/mediaLists.controller.ts
@@ -1,6 +1,5 @@
 import { Body, Controller, Delete, Get, Param, Patch, Post, Query, UseGuards } from "@nestjs/common"
 import { isCuid } from "@paralleldrive/cuid2"
-import { UserDto } from "@/services/auth/dto/user.dto"
 import { AuthGuard } from "@/services/auth/guards/auth.guard"
 import { CreateMediaListDto } from "@/services/mediaLists/dto/createMediaList.dto"
 import { CreateMediaListCloneDto } from "@/services/mediaLists/dto/createMediaListClone.dto"
@@ -8,6 +7,7 @@ import { GetAllMediaListsDto } from "@/services/mediaLists/dto/getAllMediaLists.
 import { GetMedialListByIdDto } from "@/services/mediaLists/dto/getMedialListById.dto"
 import { UpdateMediaListDto } from "@/services/mediaLists/dto/updateMediaList.dto"
 import { MediaListsService } from "@/services/mediaLists/mediaLists.service"
+import { UserDto } from "@/services/users/dto/user.dto"
 import { User } from "@/services/users/user.decorator"
 import { UuidDto } from "@/shared/dto/uuid.dto"
 import { UnauthorizedError } from "@/shared/errors/core"
