+++ apps/backend/src/delivery/http/trackingData/trackingData.controller.ts
@@ -1,9 +1,9 @@
 import { Body, Controller, Param, Patch, Post, UseGuards } from "@nestjs/common"
-import { UserDto } from "@/services/auth/dto/user.dto"
 import { AuthGuard } from "@/services/auth/guards/auth.guard"
 import { BulkUpdateTrackingDataDto } from "@/services/trackingData/dto/bulkUpdateTrackingData.dto"
 import { MediaItemTrackingDataDto } from "@/services/trackingData/dto/updateTrackingData.dto"
 import { TrackingDataService } from "@/services/trackingData/trackingData.service"
+import { UserDto } from "@/services/users/dto/user.dto"
 import { User } from "@/services/users/user.decorator"
 import { UuidDto } from "@/shared/dto/uuid.dto"
