+++ apps/backend/src/main.ts
@@ -5,7 +5,6 @@ import { HttpAdapterHost, NestFactory } from "@nestjs/core"
 import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger"
 import { apiReference } from "@scalar/nestjs-api-reference"
 import * as cookieParser from "cookie-parser"
-import { Response } from "express"
 import * as session from "express-session"
 import { Logger } from "nestjs-pino"
 import { AppModule } from "@/app.module"
